+++ index.js
@@ -9,6 +9,7 @@ import cloneable from './types/cloneable.js';
 import cryptokey from './types/cryptokey.js';
 import dataview from './types/dataview.js';
 import date from './types/date.js';
+import domexception from './types/domexception.js';
 import error from './types/error.js';
 import errors from './types/errors.js';
 import file from './types/file.js';
@@ -47,10 +48,10 @@ import universal from './presets/universal.js';
 /* types */
 export {
     arraybuffer, bigintObject, bigint, blob, cloneable, cryptokey, dataview,
-    date, error, errors, file, filelist, imagebitmap, imagedata, infinity,
-    intlTypes, map, nan, negativeInfinity, negativeZero, nonbuiltinIgnore,
-    primitiveObjects, regexp, resurrectable, set, typedArraysSocketio,
-    typedArrays, undef, userObject
+    date, domexception, error, errors, file, filelist, imagebitmap, imagedata,
+    infinity, intlTypes, map, nan, negativeInfinity, negativeZero,
+    nonbuiltinIgnore, primitiveObjects, regexp, resurrectable, set,
+    typedArraysSocketio, typedArrays, undef, userObject
 };
 
 /* presets */